Resolved 7 conflicts:
- .gitignore: took theirs (upstream additions)
- ace-step-ui: merged submodule separately (9 conflicts resolved)
- api_server.py: took ours (custom endpoints for scoring, cancel, stems, etc.)
- generate_music_execute/request.py: took ours (scoring, tempo_scale, pitch_shift)
- io_audio.py: took theirs (soundfile fix)
- llm_inference.py: took ours (repetition_penalty)

Upstream's decomposed acestep/api/ modules, training_v2, Streamlit UI,
Gradio refactoring, and platform support (Jetson/XPU/macOS) all landed cleanly.

+++ acestep/core/generation/handler/init_service_loader.py
@@ -59,7 +59,7 @@ class InitServiceLoaderMixin:
                     model_checkpoint_path,
                     trust_remote_code=True,
                     attn_implementation=candidate,
-                    torch_dtype=self.dtype,
+                    dtype=self.dtype,
                 )
                 attn_implementation = candidate
                 break
